+++ index.html
@@ -234,6 +234,6 @@
       </section>
     </template>
     <script src="./js/functions.js" defer></script>
-    <script src="./js/main.js" defer></script>
+    <script src="./js/main.js" type="module" defer></script>
 </body>
 </html>
